+++ tests/xliff/dummy_django_app/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains a dummy Django app to simulate the serving of static files during tests with DEBUG=False
+"""