+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 UNRELEASED
 ----------
 
+* [ [#1271](https://github.com/digitalfabrik/integreat-cms/issues/1271) ] Fix feedback API endpoint
+* [ [#1099](https://github.com/digitalfabrik/integreat-cms/issues/1099) ] Add push content API
+* [ [#1277](https://github.com/digitalfabrik/integreat-cms/issues/1277) ] Fix change detection for XLIFF import
+* [ [#1276](https://github.com/digitalfabrik/integreat-cms/issues/1276) ] Allow importing unchanged XLIFF files
+
 
 2022.3.2
 --------